==> 00.download-and-process-annotations/00.1.explore-disease-annotations.ipynb <==
import pandas as pd

Read disease annotations

df = pd.read_csv('output/human_protein_atlas_disease_annotations.tsv.gz', sep='\t')
df.head()

df = df.assign(disease_list=lambda x: x.disease_involvement.str.split('|')).explode('disease_list')
df.head()

df.disease_list.value_counts().reset_index().query('disease_list>100')

Among the disease labels with most occurrences, `Disease variant`, `Cancer-related genes`, `FDA approved drug targets` and `Intellectual disability` are the most common. Given that there are more than 500 genes with these labels, I will drop these for label retrieval analyses. For other analyses, depending on what question we want to answer, these labels may be interesting.
