==> 00.download-and-process-annotations/01.process-corum-data.ipynb <==
import pandas as pd
import numpy as np

Process CORUM data that was downloaded from `https://github.com/broadinstitute/2022_PERISCOPE/tree/main/common_files/CORUM_humanComplexes.txt`. It is not clear where this was downloaded from.

### Read CORUM data

corum_df = (
    pd.read_csv('input/CORUM_humanComplexes.txt', sep='\t')
)
corum_df.head()

### Create `|` separated complex names for each gene

corum_refactor_df = (
    corum_df[["ComplexName", "subunits(Gene name)"]].copy()
    .rename(columns={"subunits(Gene name)": "genes"})
    .assign(Symbol = lambda x: x.genes.str.split(';'))
    .explode("Symbol")
    .reset_index(drop=True)
    .groupby("Symbol").ComplexName.apply(lambda x: '|'.join(np.unique(x)))
    .reset_index()
)

corum_refactor_df.head()

### Write to file

corum_refactor_df.to_csv('output/CORUM_proteincomplex_annotations.tsv', sep='\t', index=False)
