==> 00.download-and-process-annotations/02.0.process-viral-infection-efficiency-data.ipynb <==
import pandas as pd

batch_map = {
    "Batch1": {
        "Batch_name": "2021_04_26_Batch1",
        "Sheet": "CTG Data Batch 1 ",
    },
    "Batch2": {
        "Batch_name": "2021_05_31_Batch2",
        "Sheet": " CTG Data Batch 2",
    },
    "Batch3": {
        "Batch_name": "2021_05_10_Batch3",
        "Sheet": " CTG Data Batch 3",
    },
    "Batch4": {
        "Batch_name": "2021_05_17_Batch4",
        "Sheet": "CTG Data Batch 4",
    },
    "Batch5": {
        "Batch_name": "2021_06_07_Batch5",
        "Sheet": "CTG Data Batch 5",
    },
    "Batch6": {
        "Batch_name": "2021_06_14_Batch6",
        "Sheet": "CTG Data Batch 6",
    },
    "Batch7": {
        "Batch_name": "2021_06_21_Batch7",
        "Sheet": "CTG Data Batch 7",
    },
    "Batch8": {
        "Batch_name": "2021_07_12_Batch8",
        "Sheet": "CTG Data Batch 8",
    },
    "Batch9": {
        "Batch_name": "2021_07_26_Batch9",
        "Sheet": "CTG Data Batch 9",
    },
    "Batch10": {
        "Batch_name": "2021_08_02_Batch10",
        "Sheet": "CTG Data Batch 10",
    },
    "Batch11": {
        "Batch_name": "2021_08_09_Batch11",
        "Sheet": " CTG Data Batch 11",
    },
    "Batch12": {
        "Batch_name": "2021_08_23_Batch12",
        "Sheet": "CTG Data Batch 12",
    },
    "Batch13": {
        "Batch_name": "2021_08_30_Batch13",
        "Sheet": "CTG Data Batch 12_repeat",
    },
}

cols = [
    "Plate Name ",
    "Row ",
    "Column ",
    "Gene ID ",
    "Plus Blast ",
    "Minus Blast ",
    "% I.E",
    "Minus Avg. Background ",
]

cols_order = [
    "Batch",
    "Batch_name",
    "Plate_name",
    "Well",
    "Row",
    "Column",
    "broad_sample",
    "Plus_Blast",
    "Minus_Blast",
    "I.E_percentage",
    "Minus_Avg_Background",
]

xlsx_file = pd.ExcelFile("input/JUMP-ORF-Infection-Efficiencies.xlsx")
df = pd.DataFrame()
for batch in batch_map:
    sheet_name = batch_map[batch]["Sheet"]
    df_tmp = (
        pd.read_excel(xlsx_file, sheet_name, usecols=cols)
    )
    df_tmp["Batch"] = batch
    df_tmp["Batch_name"] = batch_map[batch]["Batch_name"]
    df = pd.concat([df, df_tmp], axis=0, ignore_index=True)

# Rename columns

df = df.rename(
    columns={
        "Plate Name ": "Plate_name",
        "Row ": "Row",
        "Column ": "Column",
        "Gene ID ": "broad_sample",
        "Plus Blast ": "Plus_Blast",
        "Minus Blast ": "Minus_Blast",
        "% I.E": "I.E_percentage",
        "Minus Avg. Background ": "Minus_Avg_Background",
    }   
)

# Add well column
df = (
    df.assign(Well = lambda x: x.apply(lambda y: f"{y['Row']}{y['Column']:02}", axis=1))
)

df[cols_order].head()

df[cols_order].to_csv('output/gpp-infection-efficiencies.csv.gz', index=False, compression='gzip')
