==> 00.download-and-process-annotations/03.process-hgnc-approved-symbol.ipynb <==
import pandas as pd

hgnc_df = pd.read_csv("input/hgnc_approved_symbol.tsv", sep="\t").drop(
    columns=["Status"]
)
hgnc_df.head()

hgnc_df = (
    hgnc_df.rename(
        columns={
            'Approved symbol': 'Approved_symbol',
            'Previous symbols': 'Previous_symbols',
            'Gene group ID': 'Gene_group_ID',
            'NCBI Gene ID': 'NCBI_Gene_ID'
        }
    )
)

hgnc_df["NCBI_Gene_ID"] = hgnc_df["NCBI_Gene_ID"].astype("Int64")
hgnc_df["NCBI_Gene_ID"] = hgnc_df["NCBI_Gene_ID"].astype("str")

hgnc_df.head()

# Process Previous symbols

hgnc_df["Previous_symbols"] = hgnc_df["Previous_symbols"].str.split(", ").str.join('|')
hgnc_df.head()

test = hgnc_df.Approved_symbol.value_counts().reset_index().query('Approved_symbol > 1').rename(columns={'index': 'Approved_symbol', 'Approved_symbol': 'count'})

# Drop duplicates
hgnc_df = hgnc_df.sort_values(by='Approved_symbol', ascending=True).drop_duplicates(subset='Approved_symbol', keep='first')

hgnc_df.to_csv('output/hgnc_approved_symbol_processed.tsv', sep='\t', index=False)
