==> 00.download-and-process-annotations/04.process-ncbi-dataset.ipynb <==
import pandas as pd

ncbi_df = pd.read_csv("input/ncbi_dataset.tsv", sep="\t")
ncbi_df.head()

ncbi_df = (
    ncbi_df.rename(
        columns={
            "NCBI GeneID": "NCBI_Gene_ID",
            "Symbol": "Approved_Symbol",
        }
    )
)[["NCBI_Gene_ID", "Approved_Symbol"]]

ncbi_df.to_csv("output/ncbi_dataset_processed.tsv", sep="\t", index=False)
