==> 00.download-and-process-annotations/05.0.fix_gene_names.ipynb <==
import pandas as pd
import numpy as np

# HGNC gene name file

hgnc_df = pd.read_csv(
    "output/hgnc_approved_symbol_processed.tsv", sep='\t', dtype="string"
).rename(columns={"Approved_symbol": "Metadata_Approved_Symbol"})
hgnc_df.head()

# Create dataframes

approved_symbol_df = hgnc_df["Metadata_Approved_Symbol"].copy().drop_duplicates()
previous_symbols_df = (
    hgnc_df.loc[hgnc_df["Previous_symbols"].notna()]
    .assign(Previous_symbols=lambda x: x["Previous_symbols"].str.split("|"))
    .explode("Previous_symbols")
)[["Metadata_Approved_Symbol", "Previous_symbols"]]

# NCBI dataset

ncbi_df = pd.read_csv(
    "output/ncbi_dataset_processed.tsv", sep='\t', dtype="string"
).rename(columns={"Approved_Symbol": "Metadata_Approved_Symbol"})

### Add approved symbols to ORF gene names

orf_metadata_df = pd.read_csv("../datasets/metadata/orf.csv.gz")
orf_metadata_df.head()

orf_metadata_df.shape

Separate negcons since they won't have Approved Symbols

negcons_df = orf_metadata_df.query("Metadata_pert_type!= 'trt'").assign(
    Metadata_Approved_Symbol=""
)
negcons_df.shape

orf_metadata_df = orf_metadata_df.query("Metadata_pert_type== 'trt'")
orf_metadata_df.shape

Separate `BAD CONSTRUCT` 

bad_construct_df = orf_metadata_df.query(
    "Metadata_broad_sample=='BAD CONSTRUCT'"
).assign(Metadata_Approved_Symbol="")
bad_construct_df.shape

orf_metadata_df = orf_metadata_df.query("Metadata_broad_sample!= 'BAD CONSTRUCT'")
orf_metadata_df.shape

Separate samples with `nan` in `Metadata_Symbol`

nan_df = orf_metadata_df.query("Metadata_Symbol.isna()").assign(
    Metadata_Approved_Symbol=""
)
nan_df.shape

orf_metadata_df = orf_metadata_df.query("Metadata_Symbol.notna()")
orf_metadata_df.shape

Identify those genes whose Metadata_Symbol is the same as Approved Symbol

orf_merged_approved_symbol_df = orf_metadata_df.merge(
    approved_symbol_df,
    left_on="Metadata_Symbol",
    right_on="Metadata_Approved_Symbol",
    how="left",
    indicator=True,
)

orf_metadata_df = (
    orf_merged_approved_symbol_df.query("_merge=='left_only'")
    .drop("_merge", axis=1)
    .drop("Metadata_Approved_Symbol", axis=1)
)

orf_metadata_df.shape

orf_merged_approved_symbol_df = orf_merged_approved_symbol_df.query(
    "_merge=='both'"
).drop("_merge", axis=1)

orf_merged_approved_symbol_df.shape

Check if `Previous_symbols` can match to `Metadata_Symbol` if they can't match to `Approved_Symbol`

orf_merged_previous_symbols_df = orf_metadata_df.merge(
    previous_symbols_df,
    left_on="Metadata_Symbol",
    right_on="Previous_symbols",
    how="left",
    indicator=True,
)

orf_metadata_df = orf_merged_previous_symbols_df.query("_merge=='left_only'").drop(
    "_merge", axis=1
).drop("Previous_symbols", axis=1).drop("Metadata_Approved_Symbol", axis=1)

orf_metadata_df.shape

orf_merged_previous_symbols_df = (
    orf_merged_previous_symbols_df.query("_merge=='both'")
    .drop("_merge", axis=1)
    .drop("Previous_symbols", axis=1)
)

orf_merged_previous_symbols_df.shape

For the remaining, check if `NCBI_Gene_ID` can be used to find the `Approved_Symbol`

orf_merged_ncbi_df = orf_metadata_df.merge(
    ncbi_df,
    left_on="Metadata_NCBI_Gene_ID",
    right_on="NCBI_Gene_ID",
    how="left",
    indicator=True,
)

orf_metadata_df = (
    orf_merged_ncbi_df.query("_merge=='left_only'")
    .drop("_merge", axis=1)
    .drop("NCBI_Gene_ID", axis=1)
    .drop("Metadata_Approved_Symbol", axis=1)
).assign(Metadata_Approved_Symbol="")

orf_metadata_df.shape

orf_merged_ncbi_df = (
    orf_merged_ncbi_df.query("_merge=='both'")
    .drop("_merge", axis=1)
    .drop("NCBI_Gene_ID", axis=1)
)

orf_merged_ncbi_df.shape

Most genes for which `Approved_Symbol` is `nan` are those with `LOC` in their names. The only ones that don't have `LOC` in their names are the following

orf_metadata_df.loc[~orf_metadata_df.Metadata_Symbol.str.contains("LOC")]

Combining all the dataframes

orf_df = (
    pd.concat(
        [
            orf_metadata_df,
            bad_construct_df,
            negcons_df,
            nan_df,
            orf_merged_approved_symbol_df,
            orf_merged_previous_symbols_df,
            orf_merged_ncbi_df,
        ],
        axis=0,
        ignore_index=True,
    )
    .sort_values(by="Metadata_JCP2022", ascending=True)
    .reset_index(drop=True)
)

orf_df.shape

Write to file

orf_df.to_csv("output/orf.csv.gz", index=False, compression="gzip")

### Add approved symbols to CRISPR gene names

crispr_metadata_df = pd.read_csv("../datasets/metadata/crispr.csv.gz")
crispr_metadata_df.Metadata_NCBI_Gene_ID = (
    crispr_metadata_df.Metadata_NCBI_Gene_ID.astype("Int64")
)
crispr_metadata_df.Metadata_NCBI_Gene_ID = (
    crispr_metadata_df.Metadata_NCBI_Gene_ID.astype(str)
)
crispr_metadata_df.replace('<NA>', "", inplace=True)
crispr_metadata_df.head()

crispr_metadata_df.shape

Separate negcons since they won't have Approved Symbols

negcons_df = crispr_metadata_df.query("Metadata_NCBI_Gene_ID==''").assign(
    Metadata_Approved_Symbol=""
)
negcons_df.shape

crispr_metadata_df = crispr_metadata_df.query("Metadata_NCBI_Gene_ID!=''")
crispr_metadata_df.shape

Identify those genes whose Metadata_Symbol is the same as Approved Symbol

crispr_merged_approved_symbol_df = crispr_metadata_df.merge(
    approved_symbol_df,
    left_on="Metadata_Symbol",
    right_on="Metadata_Approved_Symbol",
    how="left",
    indicator=True,
)

crispr_metadata_df = (
    crispr_merged_approved_symbol_df.query("_merge=='left_only'")
    .drop("_merge", axis=1)
    .drop("Metadata_Approved_Symbol", axis=1)
)

crispr_metadata_df.shape

crispr_merged_approved_symbol_df = (
    crispr_merged_approved_symbol_df.query("_merge=='both'")
    .drop("_merge", axis=1)
)

crispr_merged_approved_symbol_df.shape

Check if `Previous_symbols` can match to `Metadata_Symbol` if they can't match to `Approved_Symbol`

crispr_merged_previous_symbols_df = crispr_metadata_df.merge(
    previous_symbols_df,
    left_on="Metadata_Symbol",
    right_on="Previous_symbols",
    how="left",
    indicator=True,
)

crispr_metadata_df = (
    crispr_merged_previous_symbols_df.query("_merge=='left_only'")
    .drop("_merge", axis=1)
    .drop("Previous_symbols", axis=1)
    .drop("Metadata_Approved_Symbol", axis=1)
)

crispr_metadata_df.shape

crispr_merged_previous_symbols_df = (
    crispr_merged_previous_symbols_df.query("_merge=='both'")
    .drop("_merge", axis=1)
    .drop("Previous_symbols", axis=1)
)

crispr_merged_previous_symbols_df.shape

For the remaining, check if `NCBI_Gene_ID` can be used to find the `Approved_Symbol`

crispr_merged_ncbi_df = crispr_metadata_df.merge(
    ncbi_df,
    left_on="Metadata_NCBI_Gene_ID",
    right_on="NCBI_Gene_ID",
    how="left",
    indicator=True,
)

crispr_metadata_df = (
    crispr_merged_ncbi_df.query("_merge=='left_only'")
    .drop("_merge", axis=1)
    .drop("NCBI_Gene_ID", axis=1)
    .drop("Metadata_Approved_Symbol", axis=1)
).assign(Metadata_Approved_Symbol="")

crispr_metadata_df.shape

crispr_merged_ncbi_df = (
    crispr_merged_ncbi_df.query("_merge=='both'")
    .drop("_merge", axis=1)
    .drop("NCBI_Gene_ID", axis=1)
)

crispr_merged_ncbi_df.shape

Combine all the dataframes

crispr_df = (
    pd.concat(
        [
            negcons_df,
            crispr_merged_approved_symbol_df,
            crispr_merged_previous_symbols_df,
            crispr_merged_ncbi_df,
        ],
        axis=0,
        ignore_index=True,
    )
    .sort_values(by="Metadata_JCP2022", ascending=True)
    .reset_index(drop=True)
)

crispr_df.shape

Write to file

crispr_df.to_csv("output/crispr.csv.gz", index=False, compression="gzip")
