==> 00.download-and-process-annotations/07.process-homologues-data.ipynb <==
import pandas as pd
import numpy as np

homologues_df = pd.read_csv(
    "input/homologene.data.tsv",
    sep="\t",
    names=[
        "HID",
        "Taxonomy_ID",
        "Gene_ID",
        "Gene_Symbol",
        "Protein_gi",
        "Protein_accession",
    ],
).rename(columns={"Gene_Symbol": "Symbol"})
homologues_df.head()

homologues_df = homologues_df.groupby("Symbol").Taxonomy_ID.unique().reset_index()

homologues_df = homologues_df.assign(
    human=lambda x: x.apply(
        lambda y: True if 9606 in y["Taxonomy_ID"] else False, axis=1
    )
).assign(HID_size=lambda x: x.apply(lambda y: len(y["Taxonomy_ID"]), axis=1))

homologues_df["Homologue"] = np.where(homologues_df["HID_size"] > 1, True, False)

homologues_df.query("human == True").reset_index(drop=True)[
    ["Symbol", "Homologue"]
].to_csv("output/homologues.tsv", sep="\t", index=False)
