==> 00.download-and-process-annotations/09.download-and-process-depmap-data.ipynb <==
import pandas as pd

Download the essentiality file (CRISPRGeneDependency.csv) from the DepMap portal (https://depmap.org/portal/download/all/?releasename=DepMap+Public+23Q4) and save it outside the repository.

dependency_df = pd.read_csv("../../CRISPRGeneDependency.csv").rename(
    columns={"Unnamed: 0": "ModelID"}
)
dependency_df.head()

This file contains many cell lines. I then filter the dataframe to just the row for U2OS. The `ModelID` for U2OS is "ACH-000364" (obtained from [CPJUMP0 design](https://github.com/jump-cellpainting/jump-cellpainting/blob/f778f3998dab15997607061a44ad438ac0f1f560/0.design-pilots/4.create-perturbation-list.Rmd#L355))

dependency_df = (
    dependency_df.query('ModelID=="ACH-000364"')
    .transpose()
    .reset_index()
    .drop(index=0)
    .rename(columns={"index": "gene", 217: "probability"})
)

dependency_df[["Symbol", "geneID"]] = dependency_df["gene"].str.split(" ", expand=True)
dependency_df = dependency_df.assign(geneID=lambda x: x["geneID"].str[1:-1])

dependency_df[["geneID", "Symbol", "probability"]].to_csv(
    "output/CRISPRGeneDependency_U2OS.csv.gz", index=False
)

Download the essentiality file (CRISPRGeneEffect.csv) from the DepMap portal (https://depmap.org/portal/download/all/?releasename=DepMap+Public+23Q4) and save it outside the repository.

effect_df = pd.read_csv("../../CRISPRGeneEffect.csv").rename(
    columns={"Unnamed: 0": "ModelID"}
)
effect_df.head()

This file contains many cell lines. I then filter the dataframe to just the row for U2OS. The `ModelID` for U2OS is "ACH-000364" (obtained from [CPJUMP0 design](https://github.com/jump-cellpainting/jump-cellpainting/blob/f778f3998dab15997607061a44ad438ac0f1f560/0.design-pilots/4.create-perturbation-list.Rmd#L355))

effect_df = (
    effect_df.query('ModelID=="ACH-000364"')
    .transpose()
    .reset_index()
    .drop(index=0)
    .rename(columns={"index": "gene", 203: "effect"})
)

effect_df[["Symbol", "geneID"]] = effect_df["gene"].str.split(" ", expand=True)
effect_df = effect_df.assign(geneID=lambda x: x["geneID"].str[1:-1])

effect_df[["geneID", "Symbol", "effect"]].to_csv(
    "output/CRISPRGeneEffect_U2OS.csv.gz", index=False
)

Download the expression file (OmicsExpressionProteinCodingGenesTPMLogp1.csv) from the DepMap portal (https://depmap.org/portal/download/all/?releasename=DepMap+Public+23Q4) and save it outside the repository.

expression_df = pd.read_csv(
    "../../OmicsExpressionProteinCodingGenesTPMLogp1.csv"
).rename(columns={"Unnamed: 0": "ModelID"})
expression_df.head()

This file contains many cell lines. I then filter the dataframe to just the row for U2OS. The `ModelID` for U2OS is "ACH-000364" (obtained from [CPJUMP0 design](https://github.com/jump-cellpainting/jump-cellpainting/blob/f778f3998dab15997607061a44ad438ac0f1f560/0.design-pilots/4.create-perturbation-list.Rmd#L355))

expression_df = (
    expression_df.query('ModelID=="ACH-000364"')
    .transpose()
    .reset_index()
    .drop(index=0)
    .rename(columns={"index": "gene", 881: "TPM"})
)

expression_df[["Symbol", "geneID"]] = expression_df["gene"].str.split(" ", expand=True)
expression_df = expression_df.assign(geneID=lambda x: x["geneID"].str[1:-1])

# Setting the geneID of RLN2 to 6019 which is the correct geneID (https://www.ncbi.nlm.nih.gov/gene/6019)
expression_df.loc[expression_df["Symbol"] == "RLN2", "geneID"] = "6019"

# drop duplicates

expression_df = expression_df.drop_duplicates(subset="geneID", keep="first")

expression_df[["geneID", "Symbol", "TPM"]].to_csv(
    "output/OmicsExpressionProteinCodingGenesTPMLogp1_U2OS.csv.gz", index=False
)
