==> 00.download-and-process-annotations/10.process-chromosome-location.ipynb <==
import pandas as pd 
import re

chromosome_df = (
    pd.read_csv("input/gene_chromosome_location.tsv", sep="\t")
    .rename(
        columns={
            "Approved symbol": "Approved_symbol",
            "Chromosome": "Chromosome",
            "NCBI Gene ID": "NCBI_Gene_ID",
        }
    )
    .dropna(subset=["Chromosome", "NCBI_Gene_ID"])
    .query("Chromosome!='reserved'")
    .reset_index(drop=True)
)

chromosome_df["Chromosome"] = chromosome_df["Chromosome"].apply(
    lambda x: "12" if x == "12 alternate reference locus" else x
)

chromosome_df["NCBI_Gene_ID"] = chromosome_df["NCBI_Gene_ID"].astype(int)

def extract_chromosome_info(location):
    match = re.match(r"([XY]|\d+)([pq])([\d.]+)", location)  # Updated regex
    if match:
        return {
        "chromosome": match.group(1),
        "arm": match.group(2),
        "region": match.group(3),
        }
    else:
        return None

chromosome_df["chromosome_info"] = chromosome_df["Chromosome"].apply(
    extract_chromosome_info
)
chromosome_df["Chromosome_number"] = chromosome_df["chromosome_info"].apply(
    lambda x: x["chromosome"] if x else None
)
chromosome_df["Arm_name"] = chromosome_df["chromosome_info"].apply(
    lambda x: x["arm"] if x else None
)
chromosome_df["Region_name"] = chromosome_df["chromosome_info"].apply(
    lambda x: x["region"] if x else None
)
chromosome_df["Chromosome_arm"] = chromosome_df["chromosome_info"].apply(
    lambda x: x["chromosome"]+x["arm"] if x else None
)

chromosome_df.drop(columns=['chromosome_info'], inplace=True)

chromosome_df.dropna(subset=["Chromosome_number"], inplace=True)

chromosome_df.head()

chromosome_df.to_csv("output/gene_chromosome_location_processed.tsv", sep="\t", index=False)
